+++ packages/fluentui/react-northstar/src/themes/teams/types.ts
@@ -10,6 +10,7 @@ import {
 import { AccordionContentStylesProps } from '../../components/Accordion/AccordionContent';
 import { AccordionStylesProps } from '../../components/Accordion/Accordion';
 import { AccordionTitleStylesProps } from '../../components/Accordion/AccordionTitle';
+import { CarouselNavigationItemStylesProps } from '../../components/Carousel/CarouselNavigationItem';
 import { AlertStylesProps } from '../../components/Alert/Alert';
 import { AnimationProps } from '../../components/Animation/Animation';
 import { AttachmentProps } from '../../components/Attachment/Attachment';
@@ -84,6 +85,8 @@ import { CardBodyStylesProps } from '../../components/Card/CardBody';
 import { CardFooterStylesProps } from '../../components/Card/CardFooter';
 import { SvgIconStylesProps } from '../../components/SvgIcon/SvgIcon';
 import { SplitButtonStylesProps } from '../../components/SplitButton/SplitButton';
+import { CarouselNavigationStylesProps } from '../../components/Carousel/CarouselNavigation';
+import { CarouselItemStylesProps } from '../../components/Carousel/CarouselItem';
 
 export type TeamsThemeStylesProps = {
   Accordion: AccordionStylesProps;
@@ -101,9 +104,11 @@ export type TeamsThemeStylesProps = {
   Box: BoxStylesProps;
   ButtonContent: ButtonContentStylesProps;
   ButtonGroup: ButtonGroupProps;
+  CarouselItem: CarouselItemStylesProps;
   Chat: ChatStylesProps;
   ChatItem: ChatItemStylesProps;
   ChatMessage: ChatMessageStylesProps;
+  CarouselNavigation: CarouselNavigationStylesProps;
   Checkbox: CheckboxStylesProps;
   Divider: DividerProps;
   Dialog: DialogProps;
@@ -164,6 +169,7 @@ export type TeamsThemeStylesProps = {
   CardHeader: CardHeaderStylesProps;
   CardBody: CardBodyStylesProps;
   CardFooter: CardFooterStylesProps;
+  CarouselNavigationItem: CarouselNavigationItemStylesProps;
 };
 
 export type TeamsContextualColors = {